requirements.txt
Django==1.7.5
django-classy-tags==0.6.1
django-model-utils==2.2
django-templatetag-sugar==1.0
docutils==0.12
geordi==0.3
gprof2dot===dev
gunicorn==19.1.1
hurry.filesize==0.9
Unidecode==0.4.17
webkit2png==0.8.3
